pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Nikita Renjov <n.renjov@cft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
deepface = "^0.0.92"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
tf-keras = "^2.16.0"
fastapi = "^0.111.1"
pillow = "^10.4.0"
kafka-python-ng = "^2.2.2"
python-decouple = "^3.8"

[tool.poetry.scripts]
run-main = "app.main:main"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "0.18.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
